tools/requirements.txt
pdfkit==0.6.1
Pillow==7.1.2
webcolors==1.3
